+++ CHANGES.md
@@ -10,6 +10,7 @@ Improvements:
 - Improve performance of coplanar polygons (#63)
 - Improve performance of cones by ~6x (#64)
 - Improve performance of arrows (#66)
+- Improve performance of cylinders (#67)
 
 Documentation:
 - Add more examples (#59)